Fix block checkout attribution
