+++ client/blocks/upe/index.js
@@ -14,9 +14,12 @@ import {
 	expressCheckoutElementsStripeLink,
 } from 'wcstripe/blocks/express-checkout';
 import WCStripeAPI from 'wcstripe/api';
-import { getBlocksConfiguration } from 'wcstripe/blocks/utils';
+import {
+	addOrderAttributionInputsIfNotExists,
+	getBlocksConfiguration,
+	populateOrderAttributionInputs,
+} from 'wcstripe/blocks/utils';
 import './styles.scss';
-import { populateOrderAttributionInputs } from 'wcstripe/blocks/upe/populate-order-attribution-inputs';
 
 const api = new WCStripeAPI(
 	getBlocksConfiguration(),
@@ -107,6 +110,9 @@ if ( getBlocksConfiguration()?.isECEEnabled ) {
 
 // Update token labels when the checkout form is loaded.
 updateTokenLabelsWhenLoaded();
+
+// Add order attribution inputs to the page.
+addOrderAttributionInputsIfNotExists();
 
 // Populate order attribution inputs with order tracking data.
 populateOrderAttributionInputs();
