--- client/blocks/upe/populate-order-attribution-inputs.js
@@ -1,13 +1,0 @@
-/**
- * Populate order attribution inputs with order tracking data.
- *
- * @return {void}
- */
-export const populateOrderAttributionInputs = () => {
-	const orderAttribution = window?.wc_order_attribution;
-	if ( orderAttribution ) {
-		orderAttribution.setOrderTracking(
-			orderAttribution.params.allowTracking
-		);
-	}
-};
